mv tf model layers to another package
