+++ spark/dl/src/test/scala/com/intel/analytics/bigdl/nn/tf/ConstSpec.scala
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.analytics.bigdl.nn
+package com.intel.analytics.bigdl.nn.tf
 
-import org.scalatest.{FlatSpec, Matchers}
 import com.intel.analytics.bigdl.numeric.NumericFloat
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.utils.T
+import org.scalatest.{FlatSpec, Matchers}
 
 class ConstSpec extends FlatSpec with Matchers {
   "Const forward tensor" should "be correct" in {
